Add comprehensive test coverage for various applications
- Introduced unit and API tests for the outbox application, ensuring the outbox events endpoint returns filtered events correctly.
- Added extensive tests for the projects application, covering project creation, filtering, pagination, and status transitions.
- Implemented tests for the recommendations API in the recs application, validating local fallback results and reindexing permissions.
- Established user profile tests to manage favorite projects effectively.
- Created helper functions for generating test data and building XLSX files for import tests.
- Removed obsolete tests from the search and tags applications, streamlining the test suite.
- Updated pytest configuration to enhance test discovery and execution.
