+++ docs/architecture/adr/ADR-001.md
