+++ docs/architecture/contracts/event_contract.json
