+++ src/ml/tests/conftest.py
@@ -1,12 +1,18 @@
+import sys
+from pathlib import Path
+
 import pytest
 from starlette.testclient import TestClient
 
-try:
-    # Works when pytest is started from src/ml
-    from app.main import app
-except ModuleNotFoundError:
-    # Works when pytest is started from repo root
-    from src.ml.app.main import app
+SERVICE_ROOT = Path(__file__).resolve().parents[1]
+REPO_ROOT = Path(__file__).resolve().parents[3]
+
+for candidate in (SERVICE_ROOT, REPO_ROOT):
+    candidate_str = str(candidate)
+    if candidate_str not in sys.path:
+        sys.path.insert(0, candidate_str)
+
+from app.main import app  # noqa: E402
 
 
 @pytest.fixture
