+++ src/web/apps/applications/tests/api/test_transitions.py
