+++ src/web/apps/base/tests/api/test_endpoints.py
