+++ src/web/apps/projects/tests/unit/test_models.py
@@ -1,0 +1,9 @@
+from apps.projects.models import Project, ProjectSourceType, ProjectStatus
+
+
+def test_project_defaults():
+    project = Project(title="Data platform")
+
+    assert project.status == ProjectStatus.DRAFT
+    assert project.source_type == ProjectSourceType.MANUAL
+    assert project.tech_tags == []
