+++ src/web/apps/users/tests/unit/test_models.py
@@ -1,0 +1,9 @@
+from apps.users.models import UserProfile, UserRole
+
+
+def test_user_profile_defaults():
+    profile = UserProfile()
+
+    assert profile.role == UserRole.STUDENT
+    assert profile.interests == []
+    assert profile.favorite_project_ids == []
