+++ src/web/tests/README.md
@@ -1,0 +1,2 @@
+Shared `src/web/tests/` is reserved for suites that exercise multiple Django apps at once.
+App-local tests belong under `src/web/apps/<app>/tests/{api,unit}/`.
